♻️ refactor: admin item controller small refactor
