+++ app/Enums/Catalog/ItemImageType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Enums;
+namespace App\Enums\Catalog;
 
 enum ItemImageType: string
 {
