+++ app/Enums/Collaborator/CollaboratorRole.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Enums;
+namespace App\Enums\Collaborator;
 
 enum CollaboratorRole: string
 {
