+++ app/Http/Requests/Collaborator/UpdateCollaboratorRequest.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Requests\Collaborator;
 
-use App\Enums\CollaboratorRole;
+use App\Enums\Collaborator\CollaboratorRole;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
